requirements.txt
psutil
python-telegram-bot
